requirements.txt
crewai>=0.28.8
langchain-core>=0.1.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
flask>=2.3.0
flask-cors>=4.0.0
pyyaml>=6.0